refactor to get rid of cyclic dependencies
